requirements.txt
langchain_openai
langchain_core.prompts
langchain_core
streamlit
python-dotenv